services/user-notification/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pika==1.3.2
dj_database_url==0.5.0
psycopg[binary]>=3.1
